+++ spec/factories/post_hobby.rb
@@ -2,5 +2,6 @@ FactoryBot.define do
   factory :post_hobby do
     title { Faker::Lorem.characters(number:10) }
     text { Faker::Lorem.characters(number:30) }
+    user
   end
 end
