+++ spec/spec_helper.rb
@@ -13,6 +13,8 @@
 # it.
 #
 # See https://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
+require 'capybara/rspec'
+
 RSpec.configure do |config|
   config.before(:each, type: :system) do
     driven_by :rack_test
